Improve caption palette issues
The caption wouldn't always come out in the right colour when it only
appeared for some of the GIF. The palette crunching would shift it's
colour to the nearest match.

By having a caption on throughout the entire palettegen phase, its
colour is much more likely to be included in the palette. But it isn't
perfect, especially if you use different colours for different captions.