Add caption colours to palettes explicitly
Previously a caption was overlaid on the image as the palette was being
calculated. With small captions this would occasionally lead to the
caption fill colour not being used in the palette.

Now `make_gif` will calculate and save the palette, then edit it to add
caption colours, interpolated between the fill and the stroke, so that
some level of smoothing around the edges occurs.