Option to manually add colours to GIF palettes
Sometimes a GIF palette as produced by ffmpeg will come without with a
colour that's barely present in the image but is still an important
detail (eg a control panel light, neon sign, jewellery...). Increasing
the number of colours available is not always going to capture that
extra colour, so it might need to be pushed up quite a way. This will
make the GIF unnecessarily larger just to capture one or two extra
shades.

Now we up the number of colours and turn on `show = 1` in the new
`[caption]` section, in order to see the generated palette, capture
the individual colours and include them in the GIF configuration,
and drop the number of colours again, knowing that accent colour will
still be used.