List upcoming GIFs by day
I wanted to be able to list all the upcoming GIFs that are scheduled
on a given day (Fri = Ted Lasso day) without the noise of other days,
and it is much faster to filter at this point than to pipe the results
of the next script through grep.

Added the inverse of the "skip" option at the same time, and renamed
it to full/empty to make it more obvious what is being skipped.