Implement autosizing at .45mb/s
The GIF compilation process will give a hint on the maximum size the GIF
should be given this rate, until a size is coded into the config file.

I will still be putting an actual size in the GIF config, but having it
calculate the recommended size for me saves a few seconds each time I
make a new GIF.

I arrived at 0.45mb/s by a process of trying a few and it seeming about
right for quality vs file size with full motion. More static scenes and
animation will be smaller (not a problem), and scenes of nothing but
motion will have to be reduced in size, or given a larger budget to
accomodate the motion, dealer's choice.