Show unscheduled count only with empty days
After listing the next 'n' days, it would show a count of unscheduled
GIFs. Now only show that count if the listing contains any days without
scheduled content.

(The logic is that if the days are all full, then there is no need
to think about scheduling any content yet)