Create the 404 page
It is decorated with a GIF (naturally), which is chosen at random
at generation time from a list of candidates.

Candidates are indicated with a tag. Where I've added to a source
that had a collapsed tag list, I've opened it up and sorted the
list. No other changes to tags have occurred.