Allow for multiple GIFs in a day
More than one GIF can be scheduled on the same day. Change the
assumption that the publish schedule is only 7am, include an optional
afternoon GIF at 3pm.

Updates the `next` scripts to take into account an afternoon GIF,
and clarify how it is used in the usage and README.

(Also removes the 404 rebuild, because changing twice a day is enough.)

If multiple GIFs are still scheduled at the same time on the same day,
don't tweet them all immediately, and improve the tweet text if they are
from the same source.