Upgrade Pillow to 8.2.0
The test images need to change as Pillow 8 changed text handling in
two ways: first, introducing the anchor option to determine where the
text is placed compared to the offset; second, fixed a bug where the
text would incorrectly be offset by the same amount as the stroke
width.

This, unfortunately, also breaks the ability to rebuild any older
GIFs without them changing as I cannot find a setting where the captions
aren't altered in the rendering.