Fix problems that were going to occur
The site might build just fine right now, but that is no guarantee
that it will later. I occasionally make errors in config files that
only get found on the day the GIF is supposed to be "published",
as this site is setup to exclude future dated content.

Adds a Makefile test target for building the *entire* site, and
fixes the errors that this found.