Adds option to list RGB values of chosen palette
When a low number of colours is being used in a GIF, sometimes a spot
colour will be missing. When I need a specific colour in a GIF, I turn
up the available colours to 254, grab the RGB value(s), and put them
into a `palette.add` array.

It is easier to copy and paste the RGB colour value from the Terminal
to the text editor than it is to use xScope to inspect a colour in a
PNG, copy it, paste it, and reformat it to valid TOML. Especially with
macOS colour profiles getting in the way.