Don’t accidentally overwrite a GIF
It is a little too easy at times to create two different GIFs with the
same title, which would overwrite the original if it hadn't yet been
committed (think two different points in WarGames showing the DEFCON
board but absent-mindedly calling both GIFs "DEFCON").

If the GIF config already exists, prompt to ensure overwriting is what
was wanted (because sometimes, it is).