Reminder on new GIFs without any size reduction
Whilst it is not mandatory to reduce a GIFs output size, more often
than not I do, as there can be little to no noticable difference in
visual quality with a substantial byte saving.

However, sometimes when making a series of new GIFs I will forget to
experiment reducing the size before declaring it done and continuing
on. So, prompt on exit in the case where there is no loss.