Remove the timing cue code
In dc73ad2 I added an option for a timing cue in GIFs. It was never
documented in the README, only included in the default draft config of
a new GIF.

In 7e8dd21, the option was removed from the default draft config as I
was finding myself using Preview to find accurate frame timings. I have
never missed its absence, so now I kill the option entirely.