Add option to stabilise output GIFs
When making a GIF that is using content filmed hand-held (or faking it)
it can be shaky. Whilst this is a great dramatic look, it adds
considerable size to a GIF as each frame changes.

The option "stabilise" takes either a number (frames before/after the
current frame to use for stability detection), or a special keyword
`static` that tries to emulate a static camera.

The vidstab library also comes with a `tripod` option but that is not
being used as it doesn't behave well when the input cuts from one scene
to another (very common in GIFs); `static` is almost as good and does
survive scene changes.