Add Ted Lasso christmas animated short
This involves pushing back Galaxy Quest and Invader Zim so that I can
put some on Christmas day itself. Hence, the `push` script now preserves
the timestamp.

Also increase the possible publishing/tweeting slots to 3 (morning,
afternoon, and evening) so that the rest of these new GIFs can be
scheduled in the run-up to Christmas without pushing anything else.